requirements.txt
torch~=2.10.0
pandas~=3.0.0
torchvision~=0.25.0
pillow~=12.1.0
tqdm~=4.67.1
opencv-python~=4.13.0.90